# app/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " API of B2NH",
        "description": "This is the documentation of Api of B2NH",
        "contact": {
            "name": "Team of B2NH",
            "email": "martin@erpicum.net"
        },
        "license": {
            "name": "GNU General Public License v3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://b2nh-api.tintamarre.be/api/v1/"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Main"
                ],
                "summary": "Get the homepage of the API",
                "description": "Links to API resources",
                "operationId": "main",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/earthquake_events/": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Earthquake events informations",
                "description": "Get Earthquake events paginated",
                "operationId": "getEarthquakeEventsInfo",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/events/random": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get a random event",
                "operationId": "getRandomEvent",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/events/{type}/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get a specific event by type and id",
                "operationId": "getEvent",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Event type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "volcano",
                                "earthquake",
                                "tsunami"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/images/{type}/{key}": {
            "get": {
                "tags": [
                    "Image"
                ],
                "summary": "Get image URL",
                "description": "Get image url of event",
                "operationId": "getImageURL",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type of image",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "volcano"
                    },
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Key of image",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "etna"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/volcanoes_map/": {
            "get": {
                "tags": [
                    "Volcano"
                ],
                "summary": "Get Volcanoes geographic informations",
                "description": "Get Volcanoes lat/long as GeoJson",
                "operationId": "getVolcanoInfo",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/filter_map/start/{start_year}/end/{end_year}": {
            "get": {
                "tags": [
                    "Map"
                ],
                "summary": "Get events between start and end year in GeoJSON",
                "operationId": "getEventsGeojson",
                "parameters": [
                    {
                        "name": "start_year",
                        "in": "path",
                        "description": "Start year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "-5000"
                    },
                    {
                        "name": "end_year",
                        "in": "path",
                        "description": "End year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2021"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/filter_map_array/start/{start_year}/end/{end_year}": {
            "get": {
                "tags": [
                    "Map"
                ],
                "summary": "Get events between start and end year in Array",
                "operationId": "getEventsAsArray",
                "parameters": [
                    {
                        "name": "start_year",
                        "in": "path",
                        "description": "Start year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "-5000"
                    },
                    {
                        "name": "end_year",
                        "in": "path",
                        "description": "End year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2021"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/tsunami_events/": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Tsunami events informations",
                "description": "Get Tsunami events paginated",
                "operationId": "getTsunamiEventsInfo",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/volcanoes/": {
            "get": {
                "tags": [
                    "Volcano"
                ],
                "summary": "Get Volcanoes informations",
                "description": "Get Volcanoes paginated",
                "operationId": "getVolcanoesInfo",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/volcanoes/{id}": {
            "get": {
                "tags": [
                    "Volcano"
                ],
                "summary": "Get Volcano information",
                "description": "Get Volcano",
                "operationId": "getVolcanoInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the volcano",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10106"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/volcanoes_elevation/": {
            "get": {
                "tags": [
                    "Volcano"
                ],
                "summary": "Get Volcanoes elevations",
                "description": "Get Volcanoes elevations",
                "operationId": "getVolcanoInfo",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/volcano_events/": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Volcano events informations",
                "description": "Get Volcano events paginated",
                "operationId": "getVolcanoEventsInfo",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "200": {
                "description": "Success"
            },
            "403": {
                "description": "Forbidden"
            },
            "404": {
                "description": "Resource Not Found"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "main",
            "description": "Everything about"
        }
    ]
}